% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult_seeds}
\alias{adult_seeds}
\title{Adult Seeds}
\format{
A data frame with 31 rows and 3 variables
\describe {
\item{watershed}{name of watershed}
\item{order}{order}
\item{seed}{number of adults for seeding at watershed}
}
}
\source{
Adam Duarte
}
\usage{
adult_seeds
}
\description{
Adult Seeds
}
\details{
This input is still in experimental stage, and is needed by the new
iteration of the model. The process for computing these values is still unclear
}
\keyword{datasets}
