% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bypass_over}
\alias{bypass_over}
\title{Sutter and Yolo Bypass Over Topped}
\format{A 3 dimensional array: 12 by 21 by 2 [months, years, bypasses]

[ , , 1] Sutter Bypass

[ , , 2] Yolo Bypass}
\source{
Sadie Gill \email{sgill@flowwest.com}
}
\usage{
bypass_over
}
\description{
1979-2000 boolean record of the bypasses over topped based on CALSIM flows
}
\details{
For more details see:
\itemize{
  \item use this link within R \code{\link[cvpiaFlow]{bypass_overtopped}}
  \item use this \href{https://flowwest.github.io/cvpiaFlow/reference/bypass_overtopped.html}{link} if in a web browser
}
}
\keyword{datasets}
