% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bypass_over_top}
\alias{bypass_over_top}
\title{Sutter and Yolo Bypass Over Topped}
\format{A data frame with 82 rows and 3 variables:
\describe{
   \item{year}{Year 1921-2002}
   \item{sutter}{1 = bypass over topped in October or November}
   \item{yolo}{1 = bypass over topped in October or November}
}}
\source{
Mike Urkov \email{mike.urkov@gmail.com}
}
\usage{
bypass_over_top
}
\description{
Binary record of the bypasses over topped in October or November based on CALSIM flows
data compiled by Mike Urkov
}
\keyword{datasets}
