% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bypass_prop_Q}
\alias{bypass_prop_Q}
\title{Proportion Flow at Yolo and Sutter Bypasses}
\format{A 3 dimensional array: 12 by 20 by 6 [months, years, bypasses]

\itemize{
  \item [ , , 1] Sutter Bypass 1
  \item [ , , 2] Sutter Bypass 2
  \item [ , , 3] Sutter Bypass 3
  \item [ , , 4] Sutter Bypass 4
  \item [ , , 5] Yolo Bypass 1
  \item [ , , 6] Yolo Bypass 2
}}
\usage{
bypass_prop_Q
}
\description{
The proportion of Lower Sacramento River flow at each bypass weir
}
\details{
For more details see: 
\itemize{
  \item use this link within R \code{\link[cvpiaFlow]{propQbypass}}
  \item use this \href{https://flowwest.github.io/cvpiaFlow/reference/propQbypass.html}{link} if in a web browser
}
}
\keyword{datasets}
