% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_DSM_array}
\alias{create_DSM_array}
\title{takes tidy data and transforms to array data structure for model input}
\usage{
create_DSM_array(input)
}
\arguments{
\item{input}{a vector of data, length = 252 for 12 months and 20 years of data}
}
\description{
takes tidy data and transforms to array data structure for model input
}
