% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_Sit_array}
\alias{create_Sit_array}
\alias{create_SIT_array}
\title{Generate SIT Model Compatible Array}
\usage{
create_SIT_array(input)
}
\arguments{
\item{input}{a vector of data, length = 252 for 12 months and 20 years of data}
}
\value{
3 dimension array [location, month, year]
}
\description{
takes output from spread_for_array and transforms to array data structure for SIT model input
}
