% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{degday}
\alias{degday}
\title{Degree Days}
\format{a 3 dimensional array [31 watersheds, 12 months, 22 years]}
\usage{
degday
}
\description{
The monthly accumulated degree days (celsius)
}
\details{
Accumulated degree days for 1979-2000
For more details see:
\itemize{
  \item use this link within R \code{\link[cvpiaTemperature]{deg_days}}
  \item use this \href{https://flowwest.github.io/cvpiaTemperature/reference/deg_days.html}{link} if in a web browser
}
}
\keyword{datasets}
