% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dlt_divers}
\alias{dlt_divers}
\title{Delta Proportion Diverted}
\format{A 3 dimensional array: 12 by 20 by 2 (months, years, deltas)

[ , , 1] North Delta

[ , , 2] South Delta}
\source{
Additional metadata contained in the \href{https://flowwest.github.io/cvpiaFlow/}{cvpiaFlow} package
}
\usage{
dlt_divers
}
\description{
The proportion of delta inflow diverted from 1980-1999.
}
\details{
The North Delta is defined as the area west of and including the Sacramento River below Freeport to Chips Island.

The South Delta is defined as the area east of the Sacramento River below Freeport to Chips Island and the San Joaquin River
below Vernalis.
}
\keyword{datasets}
