% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dlt_inflow}
\alias{dlt_inflow}
\title{Delta Inflow}
\format{A 3 dimensional array: 12 by 20 by 2 (months, years, deltas)

[ , , 1] North Delta

[ , , 2] South Delta}
\source{
Additional metadata contained in the cvpiaFlow package
}
\usage{
dlt_inflow
}
\description{
The delta inflow in cubic meters per second from 1980-1999.
}
\details{
The North Delta is defined as the area west of and including the Sacramento River below Freeport to Chips Island.

The South Delta is defined as the area east of the Sacramento River below Freeport to Chips Island and the San Joaquin River
below Vernalis.
}
\keyword{datasets}
