% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dlt_temps}
\alias{dlt_temps}
\title{Delta Temperature}
\format{
A 3 dimensional array: 12 by 21 by 2 (months, years, deltas)

[ , , 1] North Delta

[ , , 2] South Delta
}
\usage{
dlt_temps
}
\description{
The delta temperature in °C from 1980-2000.
}
\details{
The North Delta is defined as the area west of and including the Sacramento River below Freeport to Chips Island.

The South Delta is defined as the area east of the Sacramento River below Freeport to Chips Island and the San Joaquin River
below Vernalis.

For more details see:
\itemize{
  \item use this link within R \code{\link[cvpiaTemperature]{delta_temps}}
  \item use this \href{https://flowwest.github.io/cvpiaTemperature/reference/delta_temps.html}{link} if in a web browser
}
}
\keyword{datasets}
