% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egg_temp_effect}
\alias{egg_temp_effect}
\title{Temperature Effect on Egg Mortality}
\format{A dataframe with 31 rows and 2 variables 
\describe{
  \item{watershed}{watershed within CVPIA}
  \item{mean_temp_effect}{temperature effect}
}}
\usage{
egg_temp_effect
}
\description{
The mean estimate of temperature effect used for egg to fry survival
}
\details{
\itemize{
  \item use this link within R \code{\link[cvpiaTemperature]{prop_temp_over_20_migr_cor}}
  \item use this \href{https://flowwest.github.io/cvpiaTemperature/reference/prop_temp_over_20_migr_cor.html}{link} if in a web browser
}
Was calculated by takeing the mean of dry and wet egg temp effects from previous model.
}
\keyword{datasets}
