% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{floodplain}
\alias{fr_fp}
\alias{wr_fp}
\alias{sr_fp}
\alias{st_fp}
\title{Floodplain Habitat}
\format{\itemize{
\item Fall Run and Steelhead: 3 dimensional array [31 watersheds, 12 months, 20 years]
\item Spring Run and Winter Run: 3 dimensional array [31 watersheds, 12 months, 21 years]
}}
\usage{
fr_fp

wr_fp

sr_fp

st_fp
}
\description{
The 1980-1999 total floodplain rearing habitat area in square meters for Fall Run and
Steelhead and 1980-2000 total floodplain rearing habitat area in square meters for Spring Run and
Winter Run
}
\details{
\itemize{
  \item use this link within R \code{\link[cvpiaHabitat]{set_floodplain_habitat}}
  \item use this \href{https://flowwest.github.io/cvpiaHabitat/reference/set_floodplain_habitat.html}{link} if in a web browser
}
Need to apply a suitability factor, recommend 27\%.
}
\keyword{datasets}
