% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fr_fp}
\alias{fr_fp}
\title{Floodplain Habitat - Fall Run}
\format{a 3 dimensional array [31 watersheds, 12 months, 20 years]}
\usage{
fr_fp
}
\description{
The 1980-1999 total floodplain rearing habitat area in square meters
}
\details{
\itemize{
  \item use this link within R \code{\link[cvpiaHabitat]{set_floodplain_habitat}}
  \item use this \href{https://flowwest.github.io/cvpiaHabitat/reference/set_floodplain_habitat.html}{link} if in a web browser
}
Need to apply a suitability factor, recommend 27\%.
}
\keyword{datasets}
