% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fr_known_adults}
\alias{fr_known_adults}
\title{Fall Run Known Adults (with estimates)}
\format{\describe{
  \item{order}{the order of watershed}
  \item{watershed}{name of the watershed}
  \item{1980-1999}{number of adults in the watershed, if missing esimated}
}}
\usage{
fr_known_adults
}
\description{
this dataset containts all known Fall Run adults from GrandTab as
well as estimated values for tribs with missing data on certain years.
}
\keyword{datasets}
