% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freeportQcms}
\alias{freeportQcms}
\title{Flow at Freeport}
\format{A dataframe with 12 rows and 20 variables}
\source{
Additional metadata contained in the cvpiaFlow package
}
\usage{
freeportQcms
}
\description{
The inflow at Freeport in cubic meters per second from 1980-1999.
}
\details{
Each row represents a month, each column a year from 1980-1999. 
This data is used to route fish into the delta.
}
\keyword{datasets}
