% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_temperature_data}
\alias{get_temperature_data}
\title{returns DSM array of temperature data for given year period}
\usage{
get_temperature_data(start_year, end_year, location = "watershed")
}
\arguments{
\item{start_year}{int number between 1922 and 2003}

\item{end_year}{int number between 1922 and 2003}

\item{location}{chr either 'watershed' or 'delta'}
}
\description{
returns DSM array of temperature data for given year period
}
