% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inundation_durations}
\alias{inundation_durations}
\title{Floodplain Habitat Activation Duration}
\format{
a 3 dimensional array [31 watersheds, 12 months, 20 years]
}
\usage{
inundation_durations
}
\description{
The 1980-1999 floodplain rearing habitat event duration in number of weeks
}
\details{
\itemize{
  \item use this link within R \code{\link[cvpiaHabitat]{weeks_flooded}}
  \item use this \href{https://flowwest.github.io/cvpiaHabitat/reference/weeks_flooded.html}{link} if in a web browser
}

Additional growth benefit on floodplain max out after 2 weeks, apply survival
benefit for total duration.
}
\keyword{datasets}
