% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_baseline_data.R
\name{load_baseline_data}
\alias{load_baseline_data}
\title{Load SIT Model Inputs for Years 1980-1999}
\usage{
load_baseline_data(species)
}
\arguments{
\item{species}{Call function with one of the following options to load species specific model inputs: 'fall', 'winter', 'spring', 'steelhead'}
}
\value{
list of model inputs
}
\description{
returns list of data values used by model
}
\details{
The list returned contains the following data:
\itemize{
  \item \link[=ptemp20mc]{Proportion of days in month exceeding 20°C}
  \item \link[=prop_diversion]{Proportion of flow diverted for each watershed}
  \item \link[=total_diversion]{Total flow diverted for each watershed}
  \item \link[=dlt_divers]{Proportion of flow diverted in the North and South deltas}
  \item \link[=dlt_divers_tot]{Total flow diverted in the North and South deltas}
  \item \link[=rearing_temps]{Monthly mean temperature during juvenile outmigration in each watershed}
  \item \link[=dlt_temps]{Monthly mean temperature during juvenile outmigration in the deltas}
  \item \link[=dlt_inflow]{Total inflow in the deltas}
  \item \link[=dlt_hab]{Habitat areas for the deltas}
  \item \link[=bypass_prop_Q]{Proportion of Sacramento River into bypasses}
  \item \link[=inchannel_bypass]{Instream habitat areas for each bypass}
  \item \link[=floodplain_bypass]{Floodplain habitat areas for each bypass}
  \item \link[=inundation_durations]{Floodplain habitat duration inundated for each watershed}
  \item \link[=bypass_over]{Monthly record of bypasses overtopped}
  \item \link[=degday]{Accumulated degree days for adults before spawning}
  \item \link[=returnQ]{Proportion of flows at tributary junction from natal watershed in October}
  \item \link[=upsac_flow]{Upper Sacramento flow }
  \item \link[=freeportQcms]{Flow at Georgiana slough}
  \item \link[=cross_channel_gates]{Total and proportion number of days the cross channel gates are typically closed per month}
  \item \link[=egg_temp_effect]{Temperature effect on egg mortality}
  \item \link[=misc_data]{Miscellaneous model inputs}
  \item \link[=spawn_fall]{Spawning habitat areas for each watershed}
  \item \link[=inchannel_fry_fall]{Fry instream habitat areas for each watershed}
  \item \link[=inchannel_juv_fall]{Juvenile instream habitat areas for each watershed}
  \item \link[=floodplain_fall]{Floodplain habitat areas for each watershed}
}
}
\examples{
# place these two lines within the SIT Salmon Population Model function to load the data
all_inputs <- cvpiaData::load_baseline_data('fall')
list2env(all_inputs, envir = environment())
}
