% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_SIT_data.R
\name{load_SIT_data}
\alias{load_SIT_data}
\title{Load SIT Model Inputs for Years 1980-1999}
\usage{
load_SIT_data(scenario = "baseline")
}
\arguments{
\item{scenario}{default 'baseline', move scenarios to come}
}
\value{
list of model inputs
}
\description{
returns list of data values used by model
}
