% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meanQ}
\alias{meanQ}
\title{Monthly Mean Flow}
\format{
a 3 dimensional array [31 watersheds, 12 months, 21 years]
}
\usage{
meanQ
}
\description{
The mean flow in cubic meters per second for each watershed every month of every year in the simulation (1980-2000).
}
\details{
For more details see:
\itemize{
  \item use this link within R \code{\link[cvpiaFlow]{flows_cfs}}
  \item use this \href{https://flowwest.github.io/cvpiaFlow/reference/flows_cfs.html}{link} if in a web browser
}
}
\keyword{datasets}
