% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{med_flow}
\alias{med_flow}
\title{Median Flow}
\format{a dataframe with 31 rows and 13 variables 
\describe{
  \item{watershed}{CVPIA watershed}
  \item{1}{January median flow}
  \item{2}{February median flow}
  \item{3}{March median flow}
  \item{4}{April median flow}
  \item{5}{May median flow}
  \item{6}{June median flow}
  \item{7}{July median flow}
  \item{8}{August median flow}
  \item{9}{September median flow}
  \item{10}{October median flow}
  \item{11}{November median flow}
  \item{12}{December median flow}
}}
\usage{
med_flow
}
\description{
The monthly median flows of 1980-2000 in cubic feet per second
}
\details{
For more details see:
\itemize{
  \item use this link within R \code{\link[cvpiaFlow]{flows_cfs}}
  \item use this \href{https://flowwest.github.io/cvpiaFlow/reference/flows_cfs.html}{link} if in a web browser
}
}
\keyword{datasets}
