% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_data}
\alias{misc_data}
\title{Miscellaneous Model Inputs}
\format{
a dataframe with 31 rows and 15 variables
\describe{
\item{order}{Model consistent ordering of watersheds}
\item{watershed}{Name of watershed within CVPIA}
\item{init.adult}{Intial adult escapement used for the simulation}
\item{SCDELT}{Boolean value: 1 if the watershed feeds into the South Delta}
\item{hatch.alloc}{The proportion of hatchery fish spawning}
\item{TISD}{Boolean value: 1 if the watershed can enter Sutter Bypass}
\item{YOLO}{Boolean value: 1 if the watershed can enter Yolo Bypass}
\item{p.tempMC2025}{Proportion of time from October to November in which the temperature is between 20°C and 25°C}
\item{A.HARV}{Proportion of adults harvested from golden gate until they reach their natal shed}
\item{P.scour.nst}{Estimated Probability of nests scoured}
\item{P.strand.early}{Estimated Probability of stranding early}
\item{P.strand.late}{Estimated Probability of stranding late}
\item{High.pred}{Estimated Probability of high predation during rearing}
\item{contact}{Number of contact points, estimated using PAD}
\item{prop.nat}{Proportion natural spawners}
}
}
\source{
Various \href{https://s3-us-west-2.amazonaws.com/cvpia-reference-docs/AWP+Attachment+1.pdf}{CVPIA Science Integration Team: FY18 Decision Support Model activities and FY17 priorities Memorandum}


Compiled by: James T. Peterson \email{jt.peterson@oregonstate.edu}
}
\usage{
misc_data
}
\description{
Data used within the salmon population model that have a single value per watershed.
}
\keyword{datasets}
