% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthly_reach_data}
\alias{monthly_reach_data}
\title{Watershed Data Resolved by Month}
\format{A data frame with 32472 rows and 7 variables:
\describe{
  \item{watershed}{Stream reach element}
  \item{year}{Year 1921-2003}
  \item{month}{Month January to August (1-8)}
  \item{flow}{Monthly average flow in cfs}
  \item{diversion}{Monthly average diversion in cfs}
  \item{prop_diversion}{Monthly average proportion diverted in cfs}
  \item{avg_temp}{Monthly average temperature in degrees celcius}
  }}
\source{
Mike Wright
}
\usage{
monthly_reach_data
}
\description{
A dataset containing flow, diversion, and temperature data for each stream reach
more info from Mike Wright regarding model used to generate data
Note about supplement model use for Sutter Bypass
}
\keyword{datasets}
