% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pools}
\alias{pools}
\title{Pools}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31 rows and 3 columns.
}
\usage{
pools
}
\description{
Area in square meters of pools for each watershed
}
\details{
For more details see:
\itemize{
  \item use this link within R \code{\link[cvpiaHabitat]{pools}}
  \item use this \href{https://flowwest.github.io/cvpiaHabitat/reference/pools.html}{link} if in a web browser
}
}
\keyword{datasets}
