% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prop_diversion}
\alias{prop_diversion}
\title{Proportion of Flow Diverted}
\format{a 3 dimensional array [31 watersheds, 12 months, 20 years]}
\usage{
prop_diversion
}
\description{
The proportion of flow diverted for each watershed every month of every year in the simulation (1980-1999).
}
\details{
For more details see:
\itemize{
  \item use this link within R \code{\link[cvpiaFlow]{proportion_diverted}}
  \item use this \href{https://flowwest.github.io/cvpiaFlow/reference/proportion_diverted.html}{link} if in a web browser
}
}
\keyword{datasets}
