% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prop_hatch}
\alias{prop_hatch}
\title{Proportion from Hatchery}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31 rows and 4 columns.}
\usage{
prop_hatch
}
\description{
the proportion of hatchery fish for each of the tribs.
}
\keyword{datasets}
