% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prop_pulse}
\alias{prop_pulse}
\title{Proportion Pulse Flow}
\format{a 3 dimension array [31 watersheds, 12 months, 20 years]}
\usage{
prop_pulse
}
\description{
Proportion Pulse Flow
}
\details{
This is a place holder array with zeros
Could be used for (Q_pulse - Q_base)/Q_base
}
\keyword{datasets}
