% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prop_Q_bypass}
\alias{prop_Q_bypass}
\title{Proportion Flow at Sutter and Yolo Bypasses}
\format{A data frame with 1021 rows and 4 variables:
\describe{
  \item{year}{Year 1921-2003}
  \item{month}{Month January to August (1-8)}
  \item{prop_Q_yolo}{Proportion of Lower Sacramento River flow at Yolo Bypass}
  \item{prop_Q_sutter}{Proportion of Lower Sacramento River flow at Sutter Bypass}
}}
\source{
Mike Urkov, Mike Wright
}
\usage{
prop_Q_bypass
}
\description{
The proportion of Lower Sacramento River flow at Sutter and Yolo Bypasses
Sutter data comes from Mike Urkov, Yolo data comes from Mike Wright
}
\keyword{datasets}
