% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prop_Q_yolo}
\alias{prop_Q_yolo}
\title{Proportion Flow at Yolo Bypasses}
\format{A data frame with 12 rows and 20 variables (months, years)}
\usage{
prop_Q_yolo
}
\description{
The proportion of Lower Sacramento River flow at Yolo Bypasses
}
\details{
For more details see: 
\itemize{
  \item use this link within R \code{\link[cvpiaFlow]{propQbypass}}
  \item use this \href{https://flowwest.github.io/cvpiaFlow/reference/propQbypass.html}{link} if in a web browser
}
}
\keyword{datasets}
