% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptemp20mc}
\alias{ptemp20mc}
\title{Migratory Corridor Temperature Exceedance}
\format{
A dataframe with 31 rows and 14 variables 
\describe{
  \item{order}{SIT model watershed ordering}
  \item{watershed}{watershed within CVPIA}
  \item{1}{January}
  \item{...}{}prop_temp_over_20_migr_cor
  \item{12}{December}
}
}
\usage{
ptemp20mc
}
\description{
The mean monthly proportion of temperatures above 20°C
}
\details{
\itemize{
  \item use this link within R \code{\link[cvpiaTemperature]{prop_temp_over_20_migr_cor}}
  \item use this \href{https://flowwest.github.io/cvpiaTemperature/reference/prop_temp_over_20_migr_cor.html}{link} if in a web browser
}
}
\keyword{datasets}
