% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rearing_temps}
\alias{rearing_temps}
\title{Rearing Temperature}
\format{a 3 dimensional array [31 watersheds, 12 months, 20 years]}
\usage{
rearing_temps
}
\description{
The 1980-1999 rearing temperature in °C
}
\details{
\itemize{
  \item use this link within R \code{\link[cvpiaTemperature]{juv_temp}}
  \item use this \href{https://flowwest.github.io/cvpiaTemperature/reference/juv_temp.html}{link} if in a web browser
}
}
\keyword{datasets}
