% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{return_flow}
\alias{return_flow}
\title{Return Flow}
\format{A data frame with 2542 rows and 4 variables: \describe{
  \item{watershed}{The name of the stream reach}
  \item{year}{Year 1921-2002}
  \item{retQ}{The proportion flows at tributary
  junction coming from natal watershed calculated from the average monthly
  flow in October}
  }}
\source{
James T. Peterson, Mike Urkov
}
\usage{
return_flow
}
\description{
The proportion flows at tributary junction coming from natal
  watershed.
}
\keyword{datasets}
