% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returnQ}
\alias{returnQ}
\title{Return Flow}
\format{A dataframe with 12 rows and 23 variables}
\usage{
returnQ
}
\description{
The proportion flows at tributary junction coming from natal watershed using October CALSIM II flows 
from 1979-1998.
}
\details{
Each row represents a month, each column a year from 1979-2000 

For more details see:
\itemize{
  \item use this link within R \code{\link[cvpiaFlow]{return_flow}}
  \item use this \href{https://flowwest.github.io/cvpiaFlow/reference/return_flow.html}{link} if in a web browser
}
}
\keyword{datasets}
