% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_SIT_data}
\alias{set_SIT_data}
\title{Generate SIT Model Inputs for Given 20 Year Period}
\usage{
set_SIT_data(start_year, end_year)
}
\arguments{
\item{start_year}{int number between 1922 and 2003}

\item{end_year}{int number between 1922 and 2003}
}
\value{
list of model inputs
}
\description{
returns SIT array of temperature data for given year period
}
