% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spawn_fall}
\alias{spawn_fall}
\title{Spawning Habitat - Fall Run}
\format{a 3 dimensional array [31 watersheds, 12 months, 21 years]}
\usage{
spawn_fall
}
\description{
The 1979-1999 suitable spawning habitat area in square meters
}
\details{
\itemize{
  \item use this link within R \code{\link[cvpiaHabitat]{set_spawning_habitat}}
  \item use this \href{https://flowwest.github.io/cvpiaHabitat/reference/set_spawning_habitat.html}{link} if in a web browser
}
}
\keyword{datasets}
