% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spread_for_array}
\alias{spread_for_array}
\title{Prepare data for use with create_SIT_array}
\usage{
spread_for_array(df, variable, start_year, end_year, deltas = FALSE)
}
\arguments{
\item{df}{a dataframe from CVPIAdata package}

\item{variable}{variable of interest from df ('flow', 'diversion', 'prop_diversion', 'avg_tmp')}

\item{start_year}{int number between 1922 and 2003}

\item{end_year}{int number between 1922 and 2003}

\item{deltas}{bol TRUE if retrieving data for deltas, FALSE for all other reaches}
}
\value{
dataframe with dim(number of reaches, 240)
}
\description{
takes tidy data and transforms to dataframe ready to be placed in data structure for model input
}
