% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st_fry}
\alias{st_fry}
\title{Fry Inchannel Habitat - Steelhead}
\format{a 3 dimensional array [31 watersheds, 12 months, 20 years]}
\usage{
st_fry
}
\description{
The 1980-1999 fry suitable inchannel rearing habitat area in square meters
}
\details{
\itemize{
  \item use this link within R \code{\link[cvpiaHabitat]{set_instream_habitat}}
  \item use this \href{https://flowwest.github.io/cvpiaHabitat/reference/set_instream_habitat.html}{link} if in a web browser
}
}
\keyword{datasets}
