% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempprops}
\alias{aveT20}
\alias{aveT20D}
\alias{maxT24}
\alias{maxT29}
\title{Temperature Proportions}
\format{An object of class \code{array} of dimension 31 x 12 x 20.}
\usage{
aveT20

aveT20D

maxT24

maxT29
}
\description{
a set of the temperature proportions for both tribs and the delta.
See details below for more information.
}
\details{
\itemize{
  \item{\code{aveT20} are proportions of months for which stream temperature was greater than 20째C}
  \item{\code{aveT20D} are proportions of months for which delta temperature was greater than 20째C}
  \item{\code{aveT24} probability that max temps > 24째C in a month based on average monthly temp in streams (chinook only)}
  \item{\code{aveT29} probability that max temps > 29째C in a month based on average monthly temp in streams (steelhead only)}
}
}
\keyword{datasets}
