% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{total_diversion}
\alias{total_diversion}
\title{Total Flow Diverted}
\format{a 3 dimensional array [31 watersheds, 12 months, 20 years]}
\usage{
total_diversion
}
\description{
The total flow diverted in cubic feet per second for each watershed every month of every year in the simulation (1980-1999).
}
\details{
For more details see:
\itemize{
  \item use this link within R \code{\link[cvpiaFlow]{total_diverted}}
  \item use this \href{https://flowwest.github.io/cvpiaFlow/reference/total_diverted.html}{link} if in a web browser
}
}
\keyword{datasets}
