% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upsac_flow}
\alias{upsac_flow}
\title{Upper Sacramento Flows}
\format{A dataframe with 12 rows and 21 variables}
\usage{
upsac_flow
}
\description{
Flows at bend bridge in cubic meters per second from 1980-2000.
}
\details{
Each row represents a month, each column a year from 1980-2000. 
This data is used to route fish into the delta.

For more details see:
\itemize{
  \item use this link within R \code{\link[cvpiaFlow]{upsacQ}}
  \item use this \href{https://flowwest.github.io/cvpiaFlow/reference/upsacQ.html}{link} if in a web browser
}
}
\keyword{datasets}
