% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watershed_ordering}
\alias{watershed_ordering}
\title{Ordering of Watersheds}
\format{
A data frame with 31 rows and 2 variables:
\describe{
  \item{order}{The order of the streams, used for indexing in the SIT model}
  \item{watershed}{The name of the stream reach}
}
}
\source{
James T. Peterson \email{jt.peterson@oregonstate.edu}
}
\usage{
watershed_ordering
}
\description{
Ordering of Watersheds
}
\keyword{datasets}
