% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wr_known_adults}
\alias{wr_known_adults}
\title{Spring Run Known Adults}
\format{\describe{
  \item{order}{the order of watershed}
  \item{watershed}{name of the watershed}
  \item{esimated_count}{number of adults in the watershed, if missing esimated}
  \item{1980-1999}{number of adults, no estimation}
}}
\usage{
wr_known_adults
}
\description{
this dataset containts all known Spring Run adults from GrandTab for
years 1980 to 1999.
}
\keyword{datasets}
