% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bypass_over_top}
\alias{bypass_over_top}
\title{Sutter and Yolo Bypass Over Topped}
\format{A data frame with 82 rows and 3 variables:
\describe{
   \item{month}{Integer representation of calendar month}
   \item{year}{Year 1921-2002}
   \item{tils.ove}{1 = sutter bypass over topped}
   \item{yolo.ovr}{1 = yolo bypass over topped}
}}
\source{
Mike Urkov \email{mike.urkov@gmail.com}
}
\usage{
bypass_over_top
}
\description{
Binary record of the bypasses over topped based on CALSIM flows
}
\keyword{datasets}
