% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freeportQcms}
\alias{freeportQcms}
\title{Flow at Freeport}
\format{A dataframe with 12 rows and 20 variables
Each row represents a month, each column a year from 1980-1999. 
This data is used to route fish into the delta.}
\usage{
freeportQcms
}
\description{
The inflow at Freeport in cubic meters per second from 1980-1999.
}
\details{
For more details see:
 \itemize{
  \item use this link within R \code{\link[cvpiaFlow]{freeportQ}}
  \item use this \href{https://flowwest.github.io/cvpiaFlow/reference/freeportQ.html}{link} if in a web browser
}
}
\keyword{datasets}
