% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fry}
\alias{fr_fry}
\alias{wr_fry}
\alias{sr_fry}
\alias{st_fry}
\title{Fry Inchannel Habitat}
\format{
\itemize{
\item Steelhead: 3 dimensional array [31 watersheds, 12 months, 20 years]
\item Fall Run, Spring Run and Winter Run: 3 dimensional array [31 watersheds, 12 months, 21 years]
}

An object of class \code{array} of dimension 31 x 12 x 21.

An object of class \code{array} of dimension 31 x 12 x 21.

An object of class \code{array} of dimension 31 x 12 x 21.

An object of class \code{array} of dimension 31 x 12 x 20.
}
\usage{
fr_fry

wr_fry

sr_fry

st_fry
}
\description{
The 1980-1999 fry suitable inchannel rearing habitat area in square meters for Steelhead
and 1979-2000 suitable inchannel rearing habitat area in square meters for Fall Run, Spring Run and Winter Run.
}
\details{
\itemize{
  \item use this link within R \code{\link[cvpiaHabitat]{set_instream_habitat}}
  \item use this \href{https://flowwest.github.io/cvpiaHabitat/reference/set_instream_habitat.html}{link} if in a web browser
}
}
\keyword{datasets}
