% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prop_Q_bypass}
\alias{prop_Q_bypass}
\title{Proportion Flow at Sutter and Yolo Bypasses}
\format{A data frame with 1021 rows and 4 variables:
\describe{
  \item{year}{Year 1921-2003}
  \item{month}{Month January to August (1-8)}
  \item{prop_Q_yolo}{Proportion of Lower Sacramento River flow at Yolo Bypass}
  \item{prop_Q_sutter}{Proportion of Lower Sacramento River flow at Sutter Bypass}
}}
\source{
\itemize{
  \item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
  \item \strong{Sutter Data Compiled By:} Mike Urkov \email{mike.urkov@gmail.com}
  \item \strong{Yolo Data Compiled By:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
prop_Q_bypass
}
\description{
The proportion of Lower Sacramento River flow at Sutter and Yolo Bypasses
}
\keyword{datasets}
