% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upsac_flow}
\alias{upsac_flow}
\title{Upper Sacramento Flows}
\format{A dataframe with 12 rows and 20 variables}
\source{
Additional metadata contained in the cvpiaFlow package
}
\usage{
upsac_flow
}
\description{
Flows at bend bridge in cubic meters per second from 1980-1999.
}
\details{
Each row represents a month, each column a year from 1980-1999. 
This data is used to route fish into the delta.
}
\keyword{datasets}
